feat(hero): add responsive mobile layout with scaled typography
- grid-rows-[auto_1fr] on mobile: content shrinks, map fills remaining
- md breakpoint (768px) for desktop 2-column layout
- Scaled H1: clamp(18px, 5vw, 32px)
- Scaled description: clamp(13px, 2.5vw, 15px)
- Extract HeroLayout and HeroContent components
