+++ src/components/home/hero/index.ts
@@ -1,0 +1,2 @@
+export { HeroLayout } from "./hero-layout";
+export { HeroContent } from "./hero-content";
